Enum resource_io::ResourceAction

pub enum ResourceAction {
    AddResourceEntry {
        resource_id: ResourceId,
        resource: Resource,
    },
    AddPartToResource {
        resource_id: ResourceId,
        part_id: PartId,
    },
    GetResource {
        id: ResourceId,
    },
}

Variants

AddResourceEntry

Fields

resource_id: ResourceId

resource: Resource

Adds resource entry on resource storage contract.

Requirements:

The msg::source() must be the contract admin (RMRK contract).

id can not be equal to zero.

Resource with indicated id must not exist.

Arguments:

resource_id: is a resource identifier.

resource: is a resource struct that can be Basic, Slot or Composed.

On success replies ResourceEvent::ResourceEntryAdded.

AddPartToResource

Fields

resource_id: ResourceId

part_id: PartId

Adds part ids to composed resource.

Requirements:

The msg::source() must be the contract admin (RMRK contract).

part_id must exist in the base contract.

Resource with indicated id must not exist.

Arguments:

part_id: the part id to be added to composed resource.

resource_id: the composed resource id.

On success replies ResourceEvent::PartIdAddedToResource.

GetResource

Fields

id: ResourceId

Used to check from the RMRK contract whether the resource with indicated id exists or not.

Arguments:

id: is a resource identifier.

On success replies ResourceEvent::Resource.